requirements.txt
pip==22.0.3
setuptools==60.6.0
wheel==0.37.1
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
markdown
django-filter
requests
django-test-plus
